+++ mayan/apps/converter/transformations.py
@@ -1,10 +1,9 @@
 import hashlib
 import logging
 
-from PIL import Image, ImageColor, ImageDraw, ImageFilter
+from PIL import Image, ImageColor, ImageFilter
 
 from django import forms
-from django.apps import apps
 from django.utils.encoding import force_bytes, force_text
 from django.utils.text import format_lazy
 from django.utils.translation import ugettext_lazy as _
@@ -14,6 +13,9 @@ from mayan.apps.views.http import URL
 from mayan.apps.views.widgets import ColorWidget
 
 from .layers import layer_decorations, layer_saved_transformations
+from .transformation_mixins import (
+    AssetTransformationMixin, TransformationDrawRectangleMixin
+)
 
 logger = logging.getLogger(name=__name__)
 
@@ -144,125 +146,6 @@ class BaseTransformation(metaclass=BaseTransformationType):
     def execute_on(self, image):
         self.image = image
         self.aspect = 1.0 * image.size[0] / image.size[1]
-
-
-class AssetTransformationMixin:
-    class Form(Form):
-        asset_name = forms.ChoiceField(
-            help_text=_('Asset name'), label=_('Asset'),
-            required=True
-        )
-        rotation = forms.IntegerField(
-            help_text=_(
-                'Number of degrees to rotate the image counter clockwise '
-                'around its center.'
-            ), label=_('Rotation'), required=False
-        )
-        transparency = forms.FloatField(
-            help_text=_('Opacity level of the asset in percent'),
-            label=_('Transparency'), required=False
-        )
-        zoom = forms.FloatField(
-            help_text=_('Zoom level in percent.'), label=_('Zoom'),
-            required=False
-        )
-
-        def __init__(self, *args, **kwargs):
-            super().__init__(*args, **kwargs)
-
-            def get_asset_choices():
-                queryset = apps.get_model(
-                    app_label='converter', model_name='Asset'
-                ).objects.all()
-
-                for asset in queryset.all():
-                    yield (asset.internal_name, asset)
-
-            self.fields['asset_name'].choices = get_asset_choices()
-
-    @classmethod
-    def get_arguments(cls):
-        arguments = super().get_arguments() + (
-            'asset_name', 'rotation', 'transparency', 'zoom'
-        )
-        return arguments
-
-    def _update_hash(self):
-        result = super()._update_hash()
-        asset = self.get_asset()
-        # Add the asset image hash to the transformation hash. Ensures
-        # that the content object image is updated if the asset image is
-        # updated even if the transformation itself is not updated.
-        result.update(
-            force_bytes(
-                s=asset.get_hash()
-            )
-        )
-
-        return result
-
-    def get_asset(self):
-        asset_name = getattr(self, 'asset_name', None)
-
-        Asset = apps.get_model(app_label='converter', model_name='Asset')
-
-        try:
-            asset = Asset.objects.get(internal_name=asset_name)
-        except Asset.DoesNotExist:
-            logger.error('Asset "%s" not found.', asset_name)
-            raise
-        else:
-            return asset
-
-    def get_asset_images(self):
-        try:
-            transparency = float(self.transparency or '100.0')
-        except ValueError:
-            transparency = 100
-
-        if transparency < 0:
-            transparency = 0
-        elif transparency > 100:
-            transparency = 100
-
-        try:
-            rotation = int(self.rotation or '0') % 360
-        except ValueError:
-            rotation = 0
-
-        try:
-            zoom = float(self.zoom or '100.0')
-        except ValueError:
-            zoom = 100.0
-
-        asset = self.get_asset()
-
-        image_asset = asset.get_image()
-
-        if image_asset.mode != 'RGBA':
-            image_asset.putalpha(alpha=255)
-
-        image_asset = image_asset.rotate(
-            angle=360 - rotation, resample=Image.BICUBIC,
-            expand=True
-        )
-
-        if zoom != 100.0:
-            decimal_value = zoom / 100.0
-            image_asset = image_asset.resize(
-                size=(
-                    int(image_asset.size[0] * decimal_value),
-                    int(image_asset.size[1] * decimal_value)
-                ), resample=Image.ANTIALIAS
-            )
-
-        paste_mask = image_asset.getchannel(channel='A').point(
-            lut=lambda pixel: int(pixel * transparency / 100)
-        )
-
-        return {
-            'image_asset': image_asset, 'paste_mask': paste_mask
-        }
 
 
 class TransformationAssetPaste(AssetTransformationMixin, BaseTransformation):
@@ -562,15 +445,17 @@ class TransformationCrop(BaseTransformation):
         return self.image.crop(box=(left, top, right, bottom))
 
 
-class TransformationDrawRectangle(BaseTransformation):
+class TransformationDrawRectangle(
+    TransformationDrawRectangleMixin, BaseTransformation
+):
     arguments = (
-        'left', 'top', 'right', 'bottom', 'fillcolor', 'outlinecolor',
-        'outlinewidth'
+        'left', 'top', 'right', 'bottom', 'fillcolor', 'fill_transparency',
+        'outlinecolor', 'outlinewidth'
     )
     label = _('Draw rectangle')
     name = 'draw_rectangle'
 
-    class Form(Form):
+    class Form(TransformationDrawRectangleMixin.Form):
         left = forms.IntegerField(
             help_text=_('Left side location in pixels.'), label=_('Left'),
             required=False
@@ -586,18 +471,6 @@ class TransformationDrawRectangle(BaseTransformation):
         bottom = forms.IntegerField(
             help_text=_('Bottom side location in pixels.'),
             label=_('Bottom'), required=False
-        )
-        fillcolor = forms.CharField(
-            help_text=_('Color used to fill the rectangle.'),
-            label=_('Fill color'), required=False, widget=ColorWidget()
-        )
-        outlinecolor = forms.CharField(
-            help_text=_('Color used for the outline of the rectangle.'),
-            label=_('Outline color'), required=False, widget=ColorWidget()
-        )
-        outlinewidth = forms.CharField(
-            help_text=_('Width in pixels of the rectangle outline.'),
-            label=_('Outline width'), required=False
         )
 
     def execute_on(self, *args, **kwargs):
@@ -662,39 +535,17 @@ class TransformationDrawRectangle(BaseTransformation):
         if top > bottom:
             top = bottom - 1
 
-        logger.debug(
-            'left: %f, top: %f, right: %f, bottom: %f', left, top, right,
-            bottom
-        )
+        self.bottom = bottom
+        self.left = left
+        self.top = top
+        self.right = right
 
-        fillcolor_value = getattr(self, 'fillcolor', None)
-        if fillcolor_value:
-            fill_color = ImageColor.getrgb(color=fillcolor_value)
-        else:
-            fill_color = 0
-
-        outlinecolor_value = getattr(self, 'outlinecolor', None)
-        if outlinecolor_value:
-            outline_color = ImageColor.getrgb(color=outlinecolor_value)
-        else:
-            outline_color = None
-
-        outlinewidth_value = getattr(self, 'outlinewidth', None)
-        if outlinewidth_value:
-            outline_width = int(outlinewidth_value)
-        else:
-            outline_width = 0
-
-        draw = ImageDraw.Draw(im=self.image)
-        draw.rectangle(
-            xy=(left, top, right, bottom), fill=fill_color,
-            outline=outline_color, width=outline_width
-        )
-
-        return self.image
+        return super()._execute_on(self, *args, **kwargs)
 
 
-class TransformationDrawRectanglePercent(BaseTransformation):
+class TransformationDrawRectanglePercent(
+    TransformationDrawRectangleMixin, BaseTransformation
+):
     arguments = (
         'left', 'top', 'right', 'bottom', 'fillcolor', 'fill_transparency',
         'outlinecolor', 'outlinewidth'
@@ -702,7 +553,7 @@ class TransformationDrawRectanglePercent(BaseTransformation):
     label = _('Draw rectangle (percents coordinates)')
     name = 'draw_rectangle_percent'
 
-    class Form(Form):
+    class Form(TransformationDrawRectangleMixin.Form):
         left = forms.FloatField(
             help_text=_('Left side location in percent.'),
             label=_('Left'), required=False
@@ -718,22 +569,6 @@ class TransformationDrawRectanglePercent(BaseTransformation):
         bottom = forms.FloatField(
             help_text=_('Bottom side location in percent.'),
             label=_('Bottom'), required=False
-        )
-        fillcolor = forms.CharField(
-            help_text=_('Color used to fill the rectangle.'),
-            label=_('Fill color'), required=False, widget=ColorWidget()
-        )
-        fill_transparency = forms.IntegerField(
-            help_text=_('Opacity level of the fill color in percent'),
-            label=_('Fill transparency'), required=False
-        )
-        outlinecolor = forms.CharField(
-            help_text=_('Color used for the outline of the rectangle.'),
-            label=_('Outline color'), required=False, widget=ColorWidget()
-        )
-        outlinewidth = forms.IntegerField(
-            help_text=_('Width in pixels of the rectangle outline.'),
-            label=_('Outline width'), required=False
         )
 
     def execute_on(self, *args, **kwargs):
@@ -788,44 +623,6 @@ class TransformationDrawRectanglePercent(BaseTransformation):
             bottom
         )
 
-        try:
-            fill_transparency = int(
-                getattr(self, 'fill_transparency', None) or '0'
-            )
-        except ValueError:
-            fill_transparency = 100
-        else:
-            if fill_transparency < 0:
-                fill_transparency = 0
-            elif fill_transparency > 100:
-                fill_transparency = 100
-
-        fillcolor_value = getattr(self, 'fillcolor', None)
-        if fillcolor_value:
-            fill_color = ImageColor.getrgb(color=fillcolor_value)
-        else:
-            fill_color = (0, 0, 0)
-
-        # Convert transparency to opacity. Invert intensity logic, transpose
-        # from percent to 8-bit value.
-        opacity = int(
-            (100 - fill_transparency) / 100 * 255
-        )
-
-        fill_color += (opacity,)
-
-        outlinecolor_value = getattr(self, 'outlinecolor', None)
-        if outlinecolor_value:
-            outline_color = ImageColor.getrgb(color=outlinecolor_value)
-        else:
-            outline_color = None
-
-        outlinewidth_value = getattr(self, 'outlinewidth', None)
-        if outlinewidth_value:
-            outline_width = int(outlinewidth_value)
-        else:
-            outline_width = 0
-
         left = left / 100.0 * self.image.size[0]
         top = top / 100.0 * self.image.size[1]
 
@@ -839,13 +636,12 @@ class TransformationDrawRectanglePercent(BaseTransformation):
         right = self.image.size[0] - (right / 100.0 * self.image.size[0])
         bottom = self.image.size[1] - (bottom / 100.0 * self.image.size[1])
 
-        draw = ImageDraw.Draw(im=self.image, mode='RGBA')
-        draw.rectangle(
-            xy=(left, top, right, bottom), fill=fill_color,
-            outline=outline_color, width=outline_width
-        )
+        self.bottom = bottom
+        self.left = left
+        self.top = top
+        self.right = right
 
-        return self.image
+        return super()._execute_on(self, *args, **kwargs)
 
 
 class TransformationFlip(BaseTransformation):